Add support to fpga_regtbl_wrapper for selecting a bus and multi-byte register addresses
This is required for GH platforms.

The i2c bus that the fpga's reg interface is on is 1 instead of 2.
The fpga's register table is longer than 0xFF so some registers require a 2 byte address

Testing:

Ran self test on a GH platform with test points that use these features.
Ran the test script for this script (which was already broken and required changes) and all tests passed
Ran script with and without the bus argument to confirm it defaults to i2cbus 2 when no argument is passed. 

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-6892

Signed-off-by: Joey Rainville (jrainville@nvidia.com)