add jtag self test
the jtag_test will detect the size of IR register, number of devices connected
to the jtag port over looking at the DR register.
It will in the end read the idcode of the devices found and comapre with the idcode
that is provided at the input.

Assumtions:
1. all the devices on a jtag chain are the same.
the test does not work if there are 2 different devices on the jtag chain
2. idcode provided must be lower case and it cannot have any preceding 0's
Example on ranger:
jtag_test -j /dev/jtag0 -i 0x318a0dd

Additional options:
-v for verbose
-s <N> how many miliseconds to sleep in between each jtag clock cycle
-f <N> jtag frequency in Hz. 1MHz default

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-5416