Support for independent bracket mappings provided
Previously two brackets which referred to the same input position
had to define the same mapping. For example this pattern was
allowed:

  "FPGA_SXM[1-8]_EROT_RECOV_L GPU_SXM_[0|1-8]"

while this wasn't:

  "FPGA_SXM[1-8:0-7]_EROT_RECOV_L GPU_SXM_[0|1-8]"

as it specified the mappings

| arg |   first |  second |
|     | bracket | bracket |
|-----+---------+---------|
|   1 |       0 |       1 |
|   2 |       1 |       2 |
|   3 |       2 |       3 |
|   4 |       3 |       4 |
|   5 |       4 |       5 |
|   6 |       5 |       6 |
|   7 |       6 |       7 |
|   8 |       7 |       8 |

The requirement of the same mappings for all brackets bound to
the same input was too strong. The proper requirement is for the
brackets mappings to just have the saame domains. This is the
case for both patterns above with the domain being {1, 2, 3, 4,
5, 6, 7, 8} and the latter pattern expanding to

  FPGA_SXM0_EROT_RECOV_L GPU_SXM_1
  FPGA_SXM1_EROT_RECOV_L GPU_SXM_2
  FPGA_SXM2_EROT_RECOV_L GPU_SXM_3
  FPGA_SXM3_EROT_RECOV_L GPU_SXM_4
  FPGA_SXM4_EROT_RECOV_L GPU_SXM_5
  FPGA_SXM5_EROT_RECOV_L GPU_SXM_6
  FPGA_SXM6_EROT_RECOV_L GPU_SXM_7
  FPGA_SXM7_EROT_RECOV_L GPU_SXM_8

The code was tested using extensive unit tests. Every string from
the current event_info.json was covered.

Fixes nvbug https://nvbugswb.nvidia.com/NvBugs5/SWBug.aspx?bugid=4040020&cmtNo=

Signed-off-by: Marcin Wcislo <marcin.wcislo@conclusive.pl>